+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+2020_11_21
+*  Hot Fix for Spell Settings
+
 2020_10_30
 *  Additional Bug Fixes (#278, #303, #302, #300, #304, #291, #317)